# Improve BenchmarkSerializeObject benchmark