Use sub-benchmarks with "Key=Value" naming convention to used by https://pkg.go.dev/golang.org/x/perf/cmd/benchstat.

Use ReportMetric to report response size instead of writing it to logs

/kind feature

```release-note
NONE
```

Benchmark with columns changed to be compression configuration:
* `stdlib` - standard library gzip with compression level 1.
* `disabled` - compression diabled
* `pgzip` - https://github.com/klauspost/pgzip
```
                                                   │     stdlib    │               disabled             │                pgzip                │
                                                   │    sec/op     │   sec/op     vs base               │    sec/op     vs base               │
SerializeObject/Count=1000/MediaType=Json-24         11695.5µ ± 1%   640.2µ ± 3%  -94.53% (p=0.002 n=6)   2426.4µ ± 1%  -79.25% (p=0.002 n=6)
SerializeObject/Count=1000/MediaType=Protobuf-24      8666.4µ ± 0%   501.2µ ± 2%  -94.22% (p=0.002 n=6)   2632.7µ ± 1%  -69.62% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Json-24        115.832m ± 2%   4.974m ± 1%  -95.71% (p=0.002 n=6)    8.200m ± 1%  -92.92% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Protobuf-24     85.635m ± 1%   3.671m ± 2%  -95.71% (p=0.002 n=6)    6.518m ± 1%  -92.39% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Json-24       1171.45m ± 2%   55.55m ± 5%  -95.26% (p=0.002 n=6)    63.40m ± 1%  -94.59% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Protobuf-24    859.64m ± 1%   31.75m ± 5%  -96.31% (p=0.002 n=6)    44.81m ± 8%  -94.79% (p=0.002 n=6)
geomean                                                100.2m        4.667m       -95.34%                  9.949m       -90.07%

                                                   │     stdlib    │                 disabled              │                pgzip                 │
                                                   │ responseBytes │ responseBytes  vs base                │ responseBytes  vs base               │
SerializeObject/Count=1000/MediaType=Json-24           392.1k ± 0%    4118.1k ± 0%  +950.38% (p=0.002 n=6)     324.9k ± 0%  -17.13% (p=0.002 n=6)
SerializeObject/Count=1000/MediaType=Protobuf-24       267.9k ± 0%    2644.7k ± 0%  +887.36% (p=0.002 n=6)     249.0k ± 0%   -7.05% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Json-24          3.904M ± 0%    41.181M ± 0%  +954.73% (p=0.002 n=6)     3.265M ± 0%  -16.37% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Protobuf-24      2.665M ± 0%    26.447M ± 0%  +892.54% (p=0.002 n=6)     2.472M ± 0%   -7.25% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Json-24         39.03M ± 0%    411.81M ± 0%  +955.22% (p=0.002 n=6)     32.56M ± 0%  -16.56% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Protobuf-24     26.65M ± 0%    264.47M ± 0%  +892.42% (p=0.002 n=6)     24.71M ± 0%   -7.29% (p=0.002 n=6)
geomean                                                3.230M          33.00M       +921.63%                   2.840M       -12.07%

                                                   │     stdlib    │                 disabled              │                 pgzip                 │
                                                   │     B/op      │     B/op       vs base                │     B/op       vs base                │
SerializeObject/Count=1000/MediaType=Json-24          1.004Mi ± 0%    3.933Mi ± 0%  +291.62% (p=0.002 n=6)    3.141Mi ± 2%  +212.73% (p=0.002 n=6)
SerializeObject/Count=1000/MediaType=Protobuf-24      1.011Mi ± 0%    2.527Mi ± 0%  +149.99% (p=0.002 n=6)    3.335Mi ± 1%  +229.93% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Json-24         8.015Mi ± 0%   39.277Mi ± 0%  +390.06% (p=0.002 n=6)   14.941Mi ± 1%   +86.41% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Protobuf-24     8.003Mi ± 0%   25.230Mi ± 0%  +215.25% (p=0.002 n=6)    9.592Mi ± 2%   +19.85% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Json-24       128.13Mi ± 0%   392.74Mi ± 0%  +206.51% (p=0.002 n=6)    91.99Mi ± 0%   -28.20% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Protobuf-24    64.09Mi ± 0%   252.23Mi ± 0%  +293.56% (p=0.002 n=6)    55.25Mi ± 0%   -13.79% (p=0.002 n=6)
geomean                                               9.009Mi         31.49Mi       +249.56%                  14.03Mi        +55.74%

                                                   │    stdlib  │               disabled            │                 pgzip                 │
                                                   │ allocs/op  │ allocs/op   vs base               │  allocs/op    vs base                 │
SerializeObject/Count=1000/MediaType=Json-24         59.00 ± 0%   46.00 ± 0%  -22.03% (p=0.002 n=6)     99.50 ± 1%    +68.64% (p=0.002 n=6)
SerializeObject/Count=1000/MediaType=Protobuf-24     59.00 ± 0%   46.00 ± 0%  -22.03% (p=0.002 n=6)     93.00 ± 0%    +57.63% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Json-24        62.00 ± 0%   46.00 ± 0%  -25.81% (p=0.002 n=6)    407.00 ± 0%   +556.45% (p=0.002 n=6)
SerializeObject/Count=10000/MediaType=Protobuf-24    62.00 ± 0%   46.00 ± 2%  -25.81% (p=0.002 n=6)    286.00 ± 1%   +361.29% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Json-24       69.00 ± 3%   46.00 ± 0%  -33.33% (p=0.002 n=6)   3239.00 ± 0%  +4594.20% (p=0.002 n=6)
SerializeObject/Count=100000/MediaType=Protobuf-24   67.00 ± 1%   46.00 ± 0%  -31.34% (p=0.002 n=6)   2109.00 ± 0%  +3047.76% (p=0.002 n=6)
geomean                                              62.89        46.00       -26.85%                   441.0        +601.28%

```

Conclusions:
* https://github.com/klauspost/pgzip brings 10x improvement in compression throughput while producing output smaller by 12%. This is offset by higher number allocations for larger payloads.